Moved Integer#multiple_of to Numeric#multiple_of
So other Numeric types can benefit
