+++ activesupport/lib/active_support/core_ext/integer.rb
@@ -1,3 +1,2 @@
-require 'active_support/core_ext/integer/multiple'
 require 'active_support/core_ext/integer/inflections'
 require 'active_support/core_ext/integer/time'
